# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kitelepítési Támogató Rendszer API",
        "description": "QR-kódos kitelepítési, regisztrációs és befogadóhelyi nyomon követő prototípus REST API-ja.\n\nAz autentikáció Laravel Sanctum SPA session-cookie alapú: előbb GET /sanctum/csrf-cookie, utána POST /api/login. A böngészőn kívüli kliensek (pl. ez a Swagger UI) ezért a `/api/login` végpontot csak akkor tudják sikeresen meghívni, ha a kliens ugyanabból az origin-ből fut, ahonnan a cookie-kat is fogadja.",
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Helyi fejlesztői szerver"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Rendszerfelhasználók listája",
                "description": "Csak admin szerepkör jogosult.",
                "operationId": "566aed8efb436ca506e0ef51a032ecd9",
                "responses": {
                    "200": {
                        "description": "Felhasználók listája"
                    },
                    "403": {
                        "description": "Nincs jogosultság"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Új rendszerfelhasználó létrehozása",
                "operationId": "be551c1d694a01c164966f58bfa77013",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "role_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "role_id": {
                                        "type": "integer"
                                    },
                                    "shelter_id": {
                                        "type": "string",
                                        "format": "uuid",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Létrehozott felhasználó"
                    },
                    "422": {
                        "description": "Validációs hiba"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            }
        },
        "/api/users/{targetUser}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Rendszerfelhasználó módosítása (szerepkör, befogadóhely, jelszó)",
                "operationId": "6db3c49523a18bebd9089209ad42add7",
                "parameters": [
                    {
                        "name": "targetUser",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Frissített felhasználó"
                    },
                    "403": {
                        "description": "Nincs jogosultság"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            }
        },
        "/api/users/{targetUser}/avatar": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Profilkép feltöltése/cseréje",
                "description": "Saját profilkép bárki feltöltheti, más felhasználóét csak admin.",
                "operationId": "ce1ca6f744f03d2f542c726433ea0931",
                "parameters": [
                    {
                        "name": "targetUser",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "avatar"
                                ],
                                "properties": {
                                    "avatar": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Frissített felhasználó, benne az avatar_url"
                    },
                    "403": {
                        "description": "Nincs jogosultság"
                    },
                    "422": {
                        "description": "Validációs hiba"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Profilkép eltávolítása",
                "operationId": "f6b8118379436238558f0eebb4e3a510",
                "parameters": [
                    {
                        "name": "targetUser",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Frissített felhasználó"
                    },
                    "403": {
                        "description": "Nincs jogosultság"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            }
        },
        "/api/roles": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Szerepkörök listája (felhasználókezelő űrlaphoz)",
                "operationId": "f48ce60eca4060f2597c20716ebe012e",
                "responses": {
                    "200": {
                        "description": "Szerepkörök listája"
                    }
                },
                "security": [
                    {
                        "sanctumSession": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctumSession": {
                "type": "apiKey",
                "description": "Laravel Sanctum SPA session-cookie autentikáció. Bejelentkezés: POST /api/login.",
                "name": "laravel_session",
                "in": "cookie"
            }
        }
    },
...
}